aria as ecosystem with subtokens
